requirements.txt.txt
flask==2.0.1
pandas==1.3.3
faiss==1.7.0
numpy==1.21.2
scikit-learn==0.24.2
sentence-transformers==2.1.0
torch==1.9.0